+++ crates/acts/src/cache/mod.rs
@@ -1,5 +1,6 @@
 #[allow(clippy::module_inception)]
 mod cache;
+mod writer;
 mod store;
 #[cfg(test)]
 mod tests;