+++ crates/acts/src/scheduler/tests/branch/if.rs
@@ -116,6 +116,8 @@ async fn sch_task_branch_if_false_else_running() {
         TaskState::Running
     );
 
+    rt.cache().flush();
+
     // check the branch state is updated to store
     let task = proc.task_by_nid("b1").first().unwrap().clone();
     let task_id = utils::Id::new(&task.pid, &task.id);